fix rate limiting issue
previously, if the rate limiter triggered on the last possible
change of a quantizer output (for example, when rotating the phone
upside down and keeping it there), the update would be swallowed
and the output would never reach the final value.

this update fixes that by using a worker thread which listens for
input changes, but can also deal with delayed updates.