extend ratelimitnode with support for gate function and constant rate
the idea is to be able to emit values at a (nearly) constant rate.
this would be useful for some types of filters which rely on data
arriving at a given rate (for example, simple moving-average
low pass filters).

in addition, we introduce a gate function which allows to define
which value will be passed downstream. this allows to take for
example the max() of all values which were passed into the
limiter while the limit was still in effect.

this will be useful to re-create the shake effect which we
currently hardcoded in the ControlView